+++ packages/providers/lib.esm/alchemy-provider.js
@@ -63,6 +63,12 @@ export class AlchemyProvider extends UrlJsonRpcProvider {
             case "kovan":
                 host = "eth-kovan.alchemyapi.io/v2/";
                 break;
+            case "matic":
+                host = "polygon-mainnet.g.alchemy.com/v2/";
+                break;
+            case "maticmum":
+                host = "polygon-mumbai.g.alchemy.com/v2/";
+                break;
             default:
                 logger.throwArgumentError("unsupported network", "network", arguments[0]);
         }