+++ packages/tests/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "tests/5.4.1";
+export declare const version = "tests/5.4.2";
 //# sourceMappingURL=_version.d.ts.map