+++ packages/tests/lib/test-providers.js
@@ -772,15 +772,6 @@ Object.keys(blockchainData).forEach(function (network) {
                 }
             });
         }); }, test, function (provider, network, test) {
-            // @TODO: Remove once Etherscan fixes whatever makes this unhappy
-            if (provider === "EtherscanProvider") {
-                if (hash === "0x55c477790b105e69e98afadf0505cbda606414b0187356137132bf24945016ce") {
-                    return true;
-                }
-                if (hash === "0xf724f1d6813f13fb523c5f6af6261d06d41138dd094fff723e09fb0f893f03e6") {
-                    return true;
-                }
-            }
             return false;
         });
     });
@@ -793,9 +784,7 @@ Object.keys(blockchainData).forEach(function (network) {
             name: "throws correct " + code + " error",
             networks: ["ropsten"],
             checkSkip: function (provider, network, test) {
-                // @TODO: Remove once Etherscan supports EIP-1559
-                return (code === ethers_1.ethers.utils.Logger.errors.UNPREDICTABLE_GAS_LIMIT && provider === "EtherscanProvider");
-                //return false;
+                return false;
             },
             execute: function (provider) { return __awaiter(_this, void 0, void 0, function () {
                 var value, error_1;